{{cookiecutter.package_name}}/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "project"
version = "0.0.0"
description = "Template for rapid prototyping."
authors = ["Orkun Ozoglu <ozoglu.o@phystech.edu>"]
readme = "README.md"
repository = ""

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pre-commit = "^2.21.0"
pytest = "^7.2.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[tool.pytest.ini_options]
addopts = [
    "--color=yes",
    "--durations=0",
    "--strict-markers",
    "--doctest-modules",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
log_cli = "True"
markers = [
    "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.black]
py36 = true
include = '\.pyi?$'
line_length = 79

[tool.isort]
profile = "black"
